Separate extensions better
