+++ NStackSDK/NStackSDK/Classes/Other/Constants.swift
