+++ NStackSDK/NStackSDK/Classes/Other/Extensions.swift
@@ -8,32 +8,6 @@
 
 import Foundation
 
-#if !os(macOS)
-import UIKit
-
-extension UIApplication {
-    class func safeSharedApplication() -> UIApplication? {
-        guard UIApplication.responds(to: NSSelectorFromString("sharedApplication")),
-            let unmanagedSharedApplication = UIApplication.perform(NSSelectorFromString("sharedApplication")) else {
-            return nil
-        }
-
-        return unmanagedSharedApplication.takeRetainedValue() as? UIApplication
-    }
-
-    func safeOpenURL(_ url: URL) -> Bool {
-        guard self.canOpenURL(url) else { return false }
-
-        guard let returnVal = self.perform(NSSelectorFromString("openURL:"), with: url) else {
-            return false
-        }
-
-        let value = returnVal.takeRetainedValue() as? NSNumber
-        return value?.boolValue ?? false
-    }
-}
-#endif
-
 extension FileManager {
     var documentsDirectory: URL? {
         return urls(for: .documentDirectory, in: .userDomainMask).first
